receiver/pyproject.toml
[tool.poetry]
name = "receiver"
version = "0.1.0"
description = ""
authors = ["Rate980 <2220208@ecc.ac.jp>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
pyserial = "^3.5"
pyaudio = "^0.2.14"
pigpio = "^1.78"
sliplib = "^0.6.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
